+++ TeamChess/entities/src/fr/aboucorp/teamchess/entities/model/moves/movesets/RookMoveSet.java
@@ -18,7 +18,7 @@ public class RookMoveSet extends AbstractMoveSet {
     protected SquareList getPossibleMoves(Piece piece,ChessColor turnColor) {
         SquareList validSquares = new SquareList();
         Location start = piece.getLocation();
-        for(int x = start.getX()+1; x < 8 ; x++){
+        for(float x = start.getX()+1; x < 8 ; x++){
             Square validSquare = (Square) classicBoard.getSquares().getItemByLocation(new Location(x,0,start.getZ()));
             if(validSquare != null && validSquare.getPiece() == null){
                 validSquares.add(validSquare);
@@ -30,7 +30,7 @@ public class RookMoveSet extends AbstractMoveSet {
             }
         }
 
-        for(int x = start.getX()-1 ; x >= 0 ; x--){
+        for(float x = start.getX()-1 ; x >= 0 ; x--){
             Square validSquare = (Square) classicBoard.getSquares().getItemByLocation(new Location(x,0,start.getZ()));
             if(validSquare != null && validSquare.getPiece() == null){
                 validSquares.add(validSquare);
@@ -42,7 +42,7 @@ public class RookMoveSet extends AbstractMoveSet {
             }
         }
 
-        for(int z = start.getZ()-1 ;  z >= 0; z-- ){
+        for(float z = start.getZ()-1 ;  z >= 0; z-- ){
             Square validSquare = (Square) classicBoard.getSquares().getItemByLocation(new Location(start.getX(),0,z));
             if(validSquare != null && validSquare.getPiece() == null){
                 validSquares.add(validSquare);
@@ -54,7 +54,7 @@ public class RookMoveSet extends AbstractMoveSet {
             }
         }
 
-        for(int z = start.getZ()+1 ; z < 8; z++ ){
+        for(float z = start.getZ()+1 ; z < 8; z++ ){
             Square validSquare = (Square) classicBoard.getSquares().getItemByLocation(new Location(start.getX(),0,z));
             if(validSquare != null && validSquare.getPiece() == null){
                 validSquares.add(validSquare);
